requirements.txt
fastapi
uvicorn
pydantic
google-cloud-storage
python-multipart
google-cloud-pubsub
google-cloud-documentai
